removing old collection page
