+++ packages/client/src/app/pages/portfolio/collections/index.ts
@@ -1,0 +1,2 @@
+export { CollectionsComponent } from './collections.component';
+export { CollectionsResolver } from './collections.resolver';
