+++ packages/client/src/app/pages/portfolio/index.ts
@@ -1,8 +1,6 @@
 export { PortfolioComponent } from './portfolio.component';
 export { PortHomeComponent } from './port-home/port-home.component';
 export { PortBlogPageComponent } from './port-blog-page/port-blog-page.component';
-export { PortCollectionPageComponent } from './port-collection-page/port-collection-page.component';
-
 export { WorksComponent } from './works/works.component';
 export { BlogsComponent } from './blogs/blogs.component';
 export { CollectionsComponent } from './collections/collections.component';
