--- packages/client/src/app/pages/portfolio/port-collection-page/port-collection-page.component.less
@@ -1,163 +1,0 @@
-div.collection-container {
-    width: 100%;
-
-    div.header {
-        width: 100%;
-        display: flex;
-        flex-direction: row;
-        border-bottom: 1px solid var(--site-borders);
-        align-items: center;
-        font-size: 0;
-
-        div.coll-pub {
-            i-feather {
-                stroke-width: 1;
-                position: relative;
-                top: -0.5rem;
-            }
-        }
-
-        div.coll-name {
-            flex: 1;
-
-            h2 {
-                margin-bottom: 0;
-                font-size: 4.2rem;
-            }
-        }
-
-        div.coll-desc {
-            text-align: right;
-            font-style: italic;
-
-            h3 {
-                font-size: 20px;
-                color: var(--site-text-color);
-                margin-bottom: 0;
-            }
-        }
-    }
-
-    div.works-feed {
-        display: grid;
-        grid-template-columns: repeat(auto-fill, minmax(350px, 1fr));
-        grid-gap: 1rem;
-        margin-top: 1.5rem;
-        width: 100%;
-        div.work-container {
-            display: flex;
-            flex-direction: column;
-            padding: 10px;
-            min-height: 30rem;
-            border: 1px solid var(--site-borders);
-            background-color: var(--site-accent);
-            text-align: left;
-
-            @media(max-width: 400px) {
-                width: 95%;
-            }
-    
-            div.work-header {
-                margin-bottom: 0;               
-                h1 {     
-                    display: inline;
-                    font-weight: 700;
-                    font-size: 28px;
-                    color: whitesmoke;                    
-                    a {
-                        color: whitesmoke;
-                        padding: 10px 5px 0 5px;                        
-                        &:hover {
-                            background-color: var(--site-accent-hover);
-                        }
-                    }
-                }
-                               
-        
-                h2 {
-                    display: inline-block;
-                    margin: 0 0 0 0;
-                    font-size: 14px;
-                    font-style: italic;
-                    color: whitesmoke;
-                    a {
-                        color: whitesmoke;
-                        padding: 5px 5px 2px 2px;
-                        word-break: break-all;
-                        &:hover {
-                            background-color: var(--site-accent-hover);
-                        }
-                    }
-                    span {
-                        margin: 0 0.25rem 0 0.25rem;
-                    }
-                }
-
-                div.approval {
-                    float: right;
-
-                    i-feather {
-                        margin: 0 0.15rem 0 0.25rem;
-                        stroke-width: 1;
-                        width: 16px;
-                        position: relative;
-                        top: 0.3rem;
-                    }
-                }
-        
-                div.category-genre {
-                    padding:5px 0 5px 5px;
-                    border-top: 1px solid whitesmoke;
-                    border-bottom: 1px solid whitesmoke;
-                    color: whitesmoke;
-                    font-size: 14px;
-                    span {
-                        margin: 0 0.25rem 0 0.25rem;
-                    }
-                }
-            }
-    
-            div.short-description {                
-                flex-grow: 1;
-                color: whitesmoke;
-                font-size: 14px;
-                margin-top: 1.5rem;
-    
-                img.cover-art {
-                    float: left;                    
-                    max-height: 150px;
-                    max-width: 100px;                    
-                    margin-right: 1.5rem;
-                }
-            }
-
-            div.work-stats {
-                padding-top: 5px;
-                border-top: 1px solid whitesmoke;
-                text-align: right;
-                font-size: 11px;
-                color: whitesmoke;
-                a.work-stat-button {
-                    color: whitesmoke;
-                    text-decoration: none;
-                    padding: 2px;
-                    &:hover {
-                        background-color: var(--site-accent-hover);
-                    }
-                }
-                span {
-                    margin: 0 0.25rem 0 0.25rem;
-                }
-        
-                i-feather {
-                    position: relative;
-                    width: 13px;
-                    height: 13px;
-                    stroke-width: 1;
-                    top: 0.25rem;
-                    margin-right: 0.25rem;
-                }
-            }
-        }
-    }
-}
