+++ packages/client/src/app/resolvers/index.ts
@@ -5,10 +5,8 @@ export { NewsFeedResolver } from './news-feed.resolver';
 export { BrowseFeedResolver } from './browse-feed.resolver';
 export { MyWorksResolver } from './my-works.resolver';
 export { MyBlogsResolver } from './my-blogs.resolver';
-export { MyCollectionsResolver } from './my-collections.resolver';
 export { PortWorksResolver } from './port-works.resolver';
 export { PortBlogsResolver } from './port-blogs.resolver';
-export { PortCollectionsResolver } from './port-collections.resolver';
 export { CollectionPageResolver } from './collection-page.resolver';
 export { MyStuffResolver } from './my-stuff.resolver';
 export { ViewContentResolver } from './view-content.resolver';
